refactor: fold db_errors into error.go; AGENTS file layout rule
